014_chat-with-documents/requirements.txt
pypdf
google-genai
python-dotenv
streamlit
langchain
langchain-community
langchain-google-genai
faiss-cpu
ebooklib
beautifulsoup4
